pyproject.toml
[project]
name = "ml-automation-framework"
version = "0.1.0"
description = "Python-first automation framework for standardizing data science workflows"
readme = "README.md"
requires-python = "==3.11"
license = { text = "MIT" }
authors = [{ name = "ML Platform Team" }]
keywords = ["machine-learning", "mlops", "databricks", "automation", "pipelines"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "structlog>=24.0,<25.0",

    # Data processing
    "pandas>=2.0,<3.0",
    "numpy>=1.24,<2.0",
    "pyarrow>=14.0,<16.0",

    # ML
    "scikit-learn>=1.3,<2.0",
    "xgboost==3.1.3",
    "lightgbm>=4.0,<5.0",
    "catboost>=1.2,<2.0",

    # Experiment tracking
    "mlflow>=2.10,<3.0",

    # CLI
    "typer>=0.9,<1.0",
    "rich>=13.0,<14.0",
]

[project.optional-dependencies]
databricks = [
    "databricks-sdk>=0.20,<1.0",
    "pyspark>=3.5,<4.0",
]
validation = [
    "great-expectations>=0.18,<1.0",
]
explainability = [
    "shap>=0.44,<1.0",
    "matplotlib>=3.8,<4.0",
]
tuning = [
    "optuna>=3.5,<4.0",
    "optuna-integration>=3.5,<4.0",
]
docs = [
    "mkdocs>=1.5,<2.0",
    "mkdocs-material>=9.5,<10.0",
]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-cov>=4.0,<5.0",
    "pytest-mock>=3.0,<4.0",
    "ruff>=0.3,<1.0",
    "mypy>=1.8,<2.0",
    "pre-commit>=3.0,<4.0",
    "pandas-stubs>=2.0,<3.0",
]
all-features = [
    "ml-automation-framework[validation,explainability,tuning]",
]
all = [
    "ml-automation-framework[databricks,validation,explainability,tuning,docs,dev]",
]

[project.scripts]
mlf = "ml_framework.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ml_framework"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
    "B904",   # raise from
]

[tool.ruff.lint.isort]
known-first-party = ["ml_framework"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "mlflow.*",
    "xgboost.*",
    "lightgbm.*",
    "catboost.*",
    "pyspark.*",
    "databricks.*",
    "shap.*",
    "great_expectations.*",
    "optuna.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/ml_framework"]
branch = true
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
